requirements.txt
Django==5.1.4
django-environ==0.11.2
django-split-settings==1.3.2
djangorestframework==3.15.2
drf-yasg==1.21.8